refactor: move MPS compute method to compute.rs
